pom-clean-up - move register-test-user into init
