+++ e2e/_lib/pom/init/register-test-user.ts
@@ -1,9 +1,9 @@
 import { RoutePath } from "@/lib/route-path";
 import { generateRandomGUID } from "@/lib/utilities/bark-guid";
 import { Page, expect } from "@playwright/test";
-import { getTestBirthday } from "../e2e-test-utils";
-import { initPomContext } from "../pom/init/init-pom-context";
-import { UserMyPetsPage } from "../pom/pages/user-my-pets-page";
+import { getTestBirthday } from "../../e2e-test-utils";
+import { initPomContext } from "./init-pom-context";
+import { UserMyPetsPage } from "../pages/user-my-pets-page";
 
 export async function registerTestUser(args: { page: Page }): Promise<{
   guid: string;
